move custom json read function to nhd_io
