+++ src/python_framework_v01/nhd_network_utilities_v01.py
@@ -6,7 +6,6 @@ import pandas as pd
 import zipfile
 import xarray as xr
 import network_dl
-import json
 
 
 def get_geo_file_table_rows(
@@ -719,26 +718,6 @@ def set_supernetwork_parameters(
             "layer_string": 0,
             "waterbody_parameter_file_type": "Level_Pool",
         }
-
-
-def read_custom_input_json(custom_input_file):
-    with open(custom_input_file) as json_file:
-        data = json.load(json_file)
-        supernetwork_parameters = data.get("supernetwork_parameters", None)
-        waterbody_parameters = data.get("waterbody_parameters", {})
-        forcing_parameters = data.get("forcing_parameters", {})
-        restart_parameters = data.get("restart_parameters", {})
-        output_parameters = data.get("output_parameters", {})
-        run_parameters = data.get("run_parameters", {})
-        # TODO: add error trapping for potentially missing files
-    return (
-        supernetwork_parameters,
-        waterbody_parameters,
-        forcing_parameters,
-        restart_parameters,
-        output_parameters,
-        run_parameters,
-    )
 
 
 # TODO: confirm that this function is not used, and if so, consider removing it
