+++ src/python_framework_v02/nhd_io.py
@@ -3,6 +3,7 @@ import zipfile
 import xarray as xr
 import pandas as pd
 import geopandas as gpd
+import json
 
 
 def read_netcdf(geo_file_path):
@@ -36,6 +37,26 @@ def read(geo_file_path, layer_string=None, driver_string=None):
 
 def read_mask(path, layer_string=None):
     return read_csv(path, header=None, layer_string=layer_string)
+
+
+def read_custom_input_json(custom_input_file):
+    with open(custom_input_file) as json_file:
+        data = json.load(json_file)
+        supernetwork_parameters = data.get("supernetwork_parameters", None)
+        waterbody_parameters = data.get("waterbody_parameters", {})
+        forcing_parameters = data.get("forcing_parameters", {})
+        restart_parameters = data.get("restart_parameters", {})
+        output_parameters = data.get("output_parameters", {})
+        run_parameters = data.get("run_parameters", {})
+        # TODO: add error trapping for potentially missing files
+    return (
+        supernetwork_parameters,
+        waterbody_parameters,
+        forcing_parameters,
+        restart_parameters,
+        output_parameters,
+        run_parameters,
+    )
 
 
 def replace_downstreams(data, downstream_col, terminal_code):
